+++ src/components/Loading/Loading.module.css
@@ -1,0 +1,5 @@
+.container {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
