+++ src/features/PostsWithSearch/components/BulletListWithSearch.tsx
@@ -2,8 +2,8 @@ import React from "react";
 
 import { BulletPointList } from "../../../components/Search";
 import { HighlightedTextWithSearch } from "./HighlightedTextWithSearch";
-import { Post } from "../types";
 import { searchWithPostsContextCreator } from "../hooks/hookCreator";
+import { Post } from "../../../types";
 
 const itemRenderer = (item: Post) => (
   <HighlightedTextWithSearch text={item.name} />
