+++ src/features/PostsWithSearch/index.ts
@@ -1,3 +1,2 @@
 export { searchWithPostsContextCreator } from "./hooks/hookCreator";
-export type { Post } from "./types";
 export { BulletPointListWithSearch, SearchInputWithSearch } from "./components";
