+++ src/features/PostsWithSearchWithConstate/components/BulletListWithSearch.tsx
@@ -2,8 +2,8 @@ import React from "react";
 
 import { BulletPointList } from "../../../components/Search";
 import { HighlightedTextWithSearch } from "./HighlightedTextWithSearch";
-import { Post } from "../../PostsWithSearch/types";
 import { searchWithPostsConstateCreator } from "../hooks/hookCreator";
+import { Post } from "../../../types";
 
 const itemRenderer = (item: Post) => (
   <HighlightedTextWithSearch text={item.name} />
@@ -13,6 +13,6 @@ export function BulletPointListWithSearch() {
   const { useSearchContext } = searchWithPostsConstateCreator();
   const { filteredItems } = useSearchContext();
 
-  // How do you make this work with generics? :<
+  // TODO: How do you make this work with generics? :< filteredItems is any
   return <BulletPointList items={filteredItems} itemRenderer={itemRenderer} />;
 }
