+++ src/features/PostsWithSearchWithConstate/index.ts
@@ -1,3 +1,2 @@
 export { SearchInputWithSearch, BulletPointListWithSearch } from "./components";
-export type { Post } from "./types";
 export { searchWithPostsConstateCreator } from "./hooks/hookCreator";
