--- src/features/PostsWithSearchWithConstate/types/index.ts
@@ -1,4 +1,0 @@
-export type Post = {
-  id: number;
-  name: string;
-};
