+++ src/hooks/useFetchObservableCallback.ts
@@ -1,11 +1,7 @@
 import React, { useCallback, useRef } from "react";
 import { Observable, Subscription } from "rxjs";
 
-export enum LoadingState {
-  NOT_LOADED,
-  LOADING,
-  LOADED,
-}
+import { LoadingState } from "../types";
 
 type Query<T> = () => Observable<T>;
 type ReturnType<T> = {
