+++ src/stories/01-useFetchObservable2.stories.tsx
@@ -1,7 +1,8 @@
 import React, { ChangeEvent, useCallback, useState } from "react";
 
-import useFetchObservable, { LoadingState } from "../hooks/useFetchObservable";
+import useFetchObservable  from "../hooks/useFetchObservable";
 import { getUserObservable, User } from "../api/users";
+import { LoadingState } from "../types";
 
 // https://jsonplaceholder.typicode.com/users/1
