+++ src/stories/02-useSearchWithConstate.stories.tsx
@@ -3,16 +3,11 @@ import {
   SearchInputWithSearch,
   BulletPointListWithSearch,
   searchWithPostsConstateCreator,
-  Post,
 } from "../features/PostsWithSearchWithConstate";
-
-const posts: Post[] = [
-  { id: 1, name: "My favorite post" },
-  { id: 2, name: "My least favorite post" },
-  { id: 3, name: "Not a post" },
-  { id: 4, name: "Lets toast!" },
-  { id: 5, name: "I'm at the coast" },
-];
+import useFetchObservable from "../hooks/useFetchObservable";
+import { getPosts } from "../features/PostsWithSearch/api/posts";
+import Loading from "../components/Loading";
+import { Post } from "../types";
 
 export default {
   title: "Search with constate",
@@ -20,12 +15,26 @@ export default {
 
 export const UseSearch = () => {
   const { SearchProvider } = searchWithPostsConstateCreator();
-
+  const { data: posts, error, loadingState } = useFetchObservable<Post[]>(
+    getPosts,
+    []
+  );
   return (
     <SearchProvider items={posts} itemKey="name">
-      <div>
-        <SearchInputWithSearch />
-        <BulletPointListWithSearch />
+      <div
+        style={{
+          display: "flex",
+          flexDirection: "column",
+          alignItems: "center",
+          height: 500,
+        }}
+      >
+        <div>
+          <SearchInputWithSearch />
+        </div>
+        <Loading loadingState={loadingState} error={error}>
+          <BulletPointListWithSearch />
+        </Loading>
       </div>
     </SearchProvider>
   );
