+++ src/types/index.ts
@@ -1,0 +1,10 @@
+export enum LoadingState {
+  NOT_LOADED,
+  LOADING,
+  LOADED,
+}
+
+export type Post = {
+  id: number;
+  name: string;
+};
